+++ pkg/cluster/service_map.go
@@ -4,7 +4,9 @@ import (
 	"sort"
 	"strings"
 
+	"k8s.io/client-go/pkg/api/resource"
 	"k8s.io/client-go/pkg/api/v1"
+	autoscale_v1 "k8s.io/client-go/pkg/apis/autoscaling/v1"
 	"k8s.io/client-go/pkg/apis/extensions/v1beta1"
 
 	"github.com/pearsontechnology/environment-operator/pkg/bitesize"
@@ -54,10 +56,19 @@ func (s ServiceMap) AddDeployment(deployment v1beta1.Deployment) {
 	name := deployment.Name
 
 	biteservice := s.CreateOrGet(name)
-
 	if deployment.Spec.Replicas != nil {
 		biteservice.Replicas = int(*deployment.Spec.Replicas)
 	}
+
+	if len(deployment.Spec.Template.Spec.Containers[0].Resources.Requests) != 0 {
+		cpuQuantity := new(resource.Quantity)
+		*cpuQuantity = deployment.Spec.Template.Spec.Containers[0].Resources.Requests["cpu"]
+		//		memoryQuantity := new(resource.Quantity)
+		//		*memoryQuantity = deployment.Spec.Template.Spec.Containers[0].Resources.Requests["memory"]
+		biteservice.Requests.CPU = cpuQuantity.String()
+		//		biteservice.Requests.Memory = memoryQuantity.String()
+	}
+
 	biteservice.Ssl = getLabel(deployment.ObjectMeta, "ssl") // kubeDeployment.Labels["ssl"]
 	biteservice.Version = getLabel(deployment.ObjectMeta, "version")
 	biteservice.Application = getLabel(deployment.ObjectMeta, "application")
@@ -73,6 +84,16 @@ func (s ServiceMap) AddDeployment(deployment v1beta1.Deployment) {
 
 		DeployedAt: deployment.CreationTimestamp.String(),
 	}
+}
+
+func (s ServiceMap) AddHPA(hpa autoscale_v1.HorizontalPodAutoscaler) {
+	name := hpa.Name
+
+	biteservice := s.CreateOrGet(name)
+
+	biteservice.HPA.MinReplicas = *hpa.Spec.MinReplicas
+	biteservice.HPA.MaxReplicas = hpa.Spec.MaxReplicas
+	biteservice.HPA.TargetCPUUtilizationPercentage = *hpa.Spec.TargetCPUUtilizationPercentage
 }
 
 func (s ServiceMap) AddVolumeClaim(claim v1.PersistentVolumeClaim) {