+++ CHANGELOG.md
@@ -9,16 +9,18 @@ This project adheres to [Semantic Versioning](http://semver.org/). Additionally,
 - **Fixed** for any bug fixes.
 - **Security** for any security changes or fixes for vulnerabilities.
 
-### **[0.0.3]** 
+### **[0.0.3]**
 
-#### Added 
+#### Added
 
+*  Support for configuring horizontal pod autoscaling. [[BITE-1433](https://agile-jira.pearson.com/browse/BITE-1433)]
+*  Support for configuring container CPU resource requests. [[BITE-1433](https://agile-jira.pearson.com/browse/BITE-1433)]
 *  Custom Docker registry support added for pod spec. [[BITE-1448](https://agile-jira.pearson.com/browse/BITE-1448)]
 *  Environment Operator build/release pipeline now managed by TravisCI. [[BITE-1473](https://agile-jira.pearson.com/browse/BITE-1473)]
 
 ### **[0.0.2] - 2017-01-17 [RELEASED]**
 
-* Validator command added for validation of environment.bitesize file 
+* Validator command added for validation of environment.bitesize file
 
 ### **[0.0.1] - 2017-01-17 [RELEASED]**